Updated mongodb serialization WIP: reworked entity lifetimes Began removal of ConfigureAwait in code Updated editorconfig
