+++ src/Reddoxx.Quartz.MongoDbJobStore/Models/CronTrigger.cs
@@ -1,6 +1,4 @@
-using System.Diagnostics.CodeAnalysis;
-
-using JetBrains.Annotations;
+using MongoDB.Bson;
 
 using Quartz;
 using Quartz.Impl.Triggers;
@@ -8,22 +6,57 @@ using Quartz.Spi;
 
 namespace Reddoxx.Quartz.MongoDbJobStore.Models;
 
-[UsedImplicitly(ImplicitUseTargetFlags.Members)]
 internal class CronTrigger : Trigger
 {
-    public string? CronExpression { get; set; }
+    public string? CronExpression { get; init; }
 
     /// <summary>
     /// time_zone_id
     /// </summary>
-    public required string TimeZone { get; set; }
+    public string TimeZone { get; init; }
 
 
-    public CronTrigger()
+    public CronTrigger(
+        ObjectId id,
+        string instanceName,
+        string name,
+        string group,
+        string? description,
+        DateTimeOffset? nextFireTime,
+        DateTimeOffset? previousFireTime,
+        TriggerState state,
+        DateTimeOffset startTime,
+        DateTimeOffset? endTime,
+        string? calendarName,
+        int misfireInstruction,
+        int priority,
+        JobDataMap jobDataMap,
+        JobKey jobKey,
+        string? cronExpression,
+        string timeZone
+    )
+        : base(
+            id,
+            instanceName,
+            name,
+            group,
+            description,
+            nextFireTime,
+            previousFireTime,
+            state,
+            startTime,
+            endTime,
+            calendarName,
+            misfireInstruction,
+            priority,
+            jobDataMap,
+            jobKey
+        )
     {
+        CronExpression = cronExpression;
+        TimeZone = timeZone;
     }
 
-    [SetsRequiredMembers]
     public CronTrigger(ICronTrigger trigger, TriggerState state, string instanceName)
         : base(trigger, state, instanceName)
     {
@@ -38,7 +71,9 @@ internal class CronTrigger : Trigger
             CronExpressionString = CronExpression,
             TimeZone = TimeZoneInfo.FindSystemTimeZoneById(TimeZone),
         };
+
         FillTrigger(trigger);
+
         return trigger;
     }
 }
