+++ src/Reddoxx.Quartz.MongoDbJobStore/MongoDbJobStore.cs
@@ -1,5 +1,3 @@
-using System.Globalization;
-
 using JetBrains.Annotations;
 
 using Microsoft.Extensions.DependencyInjection;
@@ -8,9 +6,7 @@ using Microsoft.Extensions.Logging;
 using MongoDB.Driver;
 
 using Quartz;
-using Quartz.Impl.AdoJobStore;
 using Quartz.Impl.Matchers;
-using Quartz.Impl.Triggers;
 using Quartz.Simpl;
 using Quartz.Spi;
 
@@ -21,13 +17,10 @@ using Reddoxx.Quartz.MongoDbJobStore.Repositories;
 using Reddoxx.Quartz.MongoDbJobStore.Serializers;
 using Reddoxx.Quartz.MongoDbJobStore.Util;
 
-using Calendar = Reddoxx.Quartz.MongoDbJobStore.Models.Calendar;
-using TriggerState = Quartz.TriggerState;
-
 namespace Reddoxx.Quartz.MongoDbJobStore;
 
 [PublicAPI]
-public class MongoDbJobStore : IJobStore
+public partial class MongoDbJobStore : IJobStore
 {
     internal static readonly SystemTextJsonObjectSerializer ObjectSerializer = new();
 
@@ -258,55 +251,6 @@ public class MongoDbJobStore : IJobStore
     }
 
 
-    #region Scheduler
-
-    public async Task SchedulerStarted(CancellationToken cancellationToken = default)
-    {
-        _logger.LogTrace("Scheduler {InstanceId}/{InstanceName} started", InstanceId, InstanceName);
-
-        if (Clustered)
-        {
-            _clusterManager = new ClusterManager(this);
-            await _clusterManager.Initialize();
-        }
-        else
-        {
-            try
-            {
-                await RecoverJobs().ConfigureAwait(false);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogError(ex, "Failure occurred during job recovery: {Message}", ex.Message);
-
-                throw new SchedulerConfigException("Failure occurred during job recovery", ex);
-            }
-        }
-
-        _misfireHandler = new MisfireHandler(this);
-        _misfireHandler.Initialize();
-        _schedulerRunning = true;
-    }
-
-    public Task SchedulerPaused(CancellationToken token = default)
-    {
-        _logger.LogTrace("Scheduler {InstanceId}/{InstanceName} paused", InstanceId, InstanceName);
-        _schedulerRunning = false;
-
-        return Task.CompletedTask;
-    }
-
-    public Task SchedulerResumed(CancellationToken token = default)
-    {
-        _logger.LogTrace("Scheduler {InstanceId}/{InstanceName} resumed", InstanceId, InstanceName);
-        _schedulerRunning = true;
-
-        return Task.CompletedTask;
-    }
-
-    #endregion
-
-
     public async Task StoreJobAndTrigger(
         IJobDetail newJob,
         IOperableTrigger newTrigger,
@@ -364,7 +308,6 @@ public class MongoDbJobStore : IJobStore
         }
     }
 
-
     public Task StoreJobsAndTriggers(
         IReadOnlyDictionary<IJobDetail, IReadOnlyCollection<ITrigger>> triggersAndJobs,
         bool replace,
@@ -421,7 +364,6 @@ public class MongoDbJobStore : IJobStore
         }
     }
 
-
     public Task<bool> RemoveJobs(IReadOnlyCollection<JobKey> jobKeys, CancellationToken token = default)
     {
         try
@@ -457,99 +399,10 @@ public class MongoDbJobStore : IJobStore
     }
 
 
-    public Task StoreTrigger(
-        IOperableTrigger newTrigger,
-        bool replaceExisting,
-        CancellationToken cancellationToken = default
-    )
-    {
-        return ExecuteInTx(
-            QuartzLockType.TriggerAccess,
-            () => StoreTriggerInternal(newTrigger, null, replaceExisting, Models.TriggerState.Waiting, false, false),
-            cancellationToken
-        );
-    }
-
-    public Task<bool> RemoveTrigger(TriggerKey triggerKey, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => RemoveTriggerInternal(triggerKey), token);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task<bool> RemoveTriggers(IReadOnlyCollection<TriggerKey> triggerKeys, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                async () =>
-                {
-                    var result = true;
-
-                    foreach (var triggerKey in triggerKeys)
-                    {
-                        result = result && await RemoveTriggerInternal(triggerKey).ConfigureAwait(false);
-                    }
-
-                    return result;
-                },
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task<bool> ReplaceTrigger(
-        TriggerKey triggerKey,
-        IOperableTrigger newTrigger,
-        CancellationToken token = default
-    )
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                () => ReplaceTriggerInternal(triggerKey, newTrigger),
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public async Task<IOperableTrigger?> RetrieveTrigger(TriggerKey triggerKey, CancellationToken token = default)
-    {
-        var result = await _triggerRepository.GetTrigger(triggerKey).ConfigureAwait(false);
-
-        return result?.GetTrigger();
-    }
-
-
-    public async Task<bool> CalendarExists(string calName, CancellationToken token = default)
-    {
-        return await _calendarRepository.CalendarExists(calName).ConfigureAwait(false);
-    }
-
-
     public async Task<bool> CheckExists(JobKey jobKey, CancellationToken token = default)
     {
         return await _jobDetailRepository.JobExists(jobKey).ConfigureAwait(false);
     }
-
 
     public async Task<bool> CheckExists(TriggerKey triggerKey, CancellationToken token = default)
     {
@@ -581,65 +434,15 @@ public class MongoDbJobStore : IJobStore
     }
 
 
-    public Task StoreCalendar(
-        string name,
-        ICalendar calendar,
-        bool replaceExisting,
-        bool updateTriggers,
-        CancellationToken cancellationToken = default
-    )
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                () => StoreCalendarInternal(name, calendar, replaceExisting, updateTriggers, cancellationToken),
-                cancellationToken
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task<bool> RemoveCalendar(string calName, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => RemoveCalendarInternal(calName), token);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public async Task<ICalendar?> RetrieveCalendar(string calName, CancellationToken token = default)
-    {
-        return await _calendarRepository.GetCalendar(calName).ConfigureAwait(false);
-    }
-
-
     public async Task<int> GetNumberOfJobs(CancellationToken token = default)
     {
         return (int)await _jobDetailRepository.GetCount().ConfigureAwait(false);
     }
 
-
     public async Task<int> GetNumberOfTriggers(CancellationToken token = default)
     {
         return (int)await _triggerRepository.GetCount().ConfigureAwait(false);
     }
-
-
-    public async Task<int> GetNumberOfCalendars(CancellationToken token = default)
-    {
-        return (int)await _calendarRepository.GetCount().ConfigureAwait(false);
-    }
-
 
     public async Task<IReadOnlyCollection<JobKey>> GetJobKeys(
         GroupMatcher<JobKey> matcher,
@@ -648,17 +451,6 @@ public class MongoDbJobStore : IJobStore
     {
         var jobsKeys = await _jobDetailRepository.GetJobsKeys(matcher).ConfigureAwait(false);
         return new HashSet<JobKey>(jobsKeys);
-    }
-
-
-    public async Task<IReadOnlyCollection<TriggerKey>> GetTriggerKeys(
-        GroupMatcher<TriggerKey> matcher,
-        CancellationToken token = default
-    )
-    {
-        var triggerKeys = await _triggerRepository.GetTriggerKeys(matcher).ConfigureAwait(false);
-
-        return new HashSet<TriggerKey>(triggerKeys);
     }
 
 
@@ -671,110 +463,6 @@ public class MongoDbJobStore : IJobStore
     public async Task<IReadOnlyCollection<string>> GetTriggerGroupNames(CancellationToken token = default)
     {
         return await _triggerRepository.GetTriggerGroupNames().ConfigureAwait(false);
-    }
-
-
-    public async Task<IReadOnlyCollection<string>> GetCalendarNames(CancellationToken token = default)
-    {
-        return await _calendarRepository.GetCalendarNames().ConfigureAwait(false);
-    }
-
-
-    public async Task<IReadOnlyCollection<IOperableTrigger>> GetTriggersForJob(
-        JobKey jobKey,
-        CancellationToken token = default
-    )
-    {
-        var triggers = await _triggerRepository.GetTriggers(jobKey).ConfigureAwait(false);
-
-        return triggers.Select(trigger => trigger.GetTrigger()).ToList();
-    }
-
-
-    public async Task<TriggerState> GetTriggerState(TriggerKey triggerKey, CancellationToken token = default)
-    {
-        var trigger = await _triggerRepository.GetTrigger(triggerKey).ConfigureAwait(false);
-
-        if (trigger == null)
-        {
-            return TriggerState.None;
-        }
-
-        return trigger.State switch
-        {
-            Models.TriggerState.Deleted => TriggerState.None,
-            Models.TriggerState.Complete => TriggerState.Complete,
-            Models.TriggerState.Paused => TriggerState.Paused,
-            Models.TriggerState.PausedBlocked => TriggerState.Paused,
-            Models.TriggerState.Error => TriggerState.Error,
-            Models.TriggerState.Blocked => TriggerState.Blocked,
-            _ => TriggerState.Normal,
-        };
-    }
-
-
-    public Task ResetTriggerFromErrorState(TriggerKey triggerKey, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                async () =>
-                {
-                    var newState = Models.TriggerState.Waiting;
-
-                    if (await _pausedTriggerGroupRepository.IsTriggerGroupPaused(triggerKey.Group))
-                    {
-                        newState = Models.TriggerState.Paused;
-                    }
-
-                    await _triggerRepository.UpdateTriggerState(triggerKey, newState, Models.TriggerState.Error);
-
-                    _logger.LogInformation(
-                        "Trigger {TriggerKey} reset from ERROR state to: {NewState}",
-                        triggerKey,
-                        newState
-                    );
-                },
-                cancellationToken
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(
-                $"Couldn't reset from error state of trigger ({triggerKey}): {ex.Message}",
-                ex
-            );
-        }
-    }
-
-
-    public Task PauseTrigger(TriggerKey triggerKey, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => PauseTriggerInternal(triggerKey), token);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task<IReadOnlyCollection<string>> PauseTriggers(
-        GroupMatcher<TriggerKey> matcher,
-        CancellationToken token = default
-    )
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => PauseTriggerGroupInternal(matcher), token);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
     }
 
 
@@ -830,35 +518,6 @@ public class MongoDbJobStore : IJobStore
                 },
                 token
             );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task ResumeTrigger(TriggerKey triggerKey, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => ResumeTriggerInternal(triggerKey), token);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task<IReadOnlyCollection<string>> ResumeTriggers(
-        GroupMatcher<TriggerKey> matcher,
-        CancellationToken token = default
-    )
-    {
-        try
-        {
-            return ExecuteInTx(QuartzLockType.TriggerAccess, () => ResumeTriggersInternal(matcher), token);
         }
         catch (Exception ex)
         {
@@ -946,7 +605,6 @@ public class MongoDbJobStore : IJobStore
         }
     }
 
-
     public Task ResumeAll(CancellationToken token = default)
     {
         try
@@ -958,187 +616,6 @@ public class MongoDbJobStore : IJobStore
             throw new JobPersistenceException(ex.Message, ex);
         }
     }
-
-
-    public Task<IReadOnlyCollection<IOperableTrigger>> AcquireNextTriggers(
-        DateTimeOffset noLaterThan,
-        int maxCount,
-        TimeSpan timeWindow,
-        CancellationToken token = default
-    )
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                () => AcquireNextTriggersInternal(noLaterThan, maxCount, timeWindow),
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public Task ReleaseAcquiredTrigger(IOperableTrigger trigger, CancellationToken token = default)
-    {
-        try
-        {
-            return ExecuteInTx(
-                QuartzLockType.TriggerAccess,
-                async () =>
-                {
-                    await _triggerRepository.UpdateTriggerState(
-                            trigger.Key,
-                            Models.TriggerState.Waiting,
-                            Models.TriggerState.Acquired
-                        )
-                        .ConfigureAwait(false);
-                    await _triggerRepository.UpdateTriggerState(
-                            trigger.Key,
-                            Models.TriggerState.Waiting,
-                            Models.TriggerState.Blocked
-                        )
-                        .ConfigureAwait(false);
-
-                    await _firedTriggerRepository.DeleteFiredTrigger(trigger.FireInstanceId).ConfigureAwait(false);
-                },
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException($"Couldn't release acquired trigger: {ex.Message}", ex);
-        }
-    }
-
-
-    public Task<IReadOnlyCollection<TriggerFiredResult>> TriggersFired(
-        IReadOnlyCollection<IOperableTrigger> triggers,
-        CancellationToken token = default
-    )
-    {
-        try
-        {
-            return ExecuteInTx<IReadOnlyCollection<TriggerFiredResult>>(
-                QuartzLockType.TriggerAccess,
-                async () =>
-                {
-                    var results = new List<TriggerFiredResult>();
-
-                    foreach (var operableTrigger in triggers)
-                    {
-                        TriggerFiredResult result;
-                        try
-                        {
-                            var bundle = await TriggerFiredInternal(operableTrigger).ConfigureAwait(false);
-                            result = new TriggerFiredResult(bundle);
-                        }
-                        catch (Exception ex)
-                        {
-                            _logger.LogError(ex, "Caught exception: {Message}", ex.Message);
-                            result = new TriggerFiredResult(ex);
-                        }
-
-                        results.Add(result);
-                    }
-
-                    return results;
-                },
-                token
-            );
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-
-    public async Task TriggeredJobComplete(
-        IOperableTrigger trigger,
-        IJobDetail jobDetail,
-        SchedulerInstruction triggerInstCode,
-        CancellationToken token = default
-    )
-    {
-        try
-        {
-            await ExecuteInTx(
-                    QuartzLockType.TriggerAccess,
-                    () => TriggeredJobCompleteInternal(trigger, jobDetail, triggerInstCode),
-                    token
-                )
-                .ConfigureAwait(false);
-
-            var sigTime = ClearAndGetSignalSchedulingChangeOnTxCompletion();
-            if (sigTime != null)
-            {
-                SignalSchedulingChangeImmediately(sigTime);
-            }
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-    internal async Task<RecoverMisfiredJobsResult> DoRecoverMisfires()
-    {
-        try
-        {
-            var misfireCount = await _triggerRepository.GetMisfireCount(MisfireTime.UtcDateTime).ConfigureAwait(false);
-
-            _logger.LogDebug("Found {MisfireCount} triggers that missed their scheduled fire-time.", misfireCount);
-
-            if (misfireCount == 0)
-            {
-                return RecoverMisfiredJobsResult.NoOp;
-            }
-
-            return await ExecuteInTx(
-                    QuartzLockType.TriggerAccess,
-                    async () => await RecoverMisfiredJobsInternal(false).ConfigureAwait(false)
-                )
-                .ConfigureAwait(false);
-        }
-        catch (Exception ex)
-        {
-            throw new JobPersistenceException(ex.Message, ex);
-        }
-    }
-
-    private Task RecoverJobs()
-    {
-        return ExecuteInTx(QuartzLockType.TriggerAccess, RecoverJobsInternal);
-    }
-
-
-    private async Task PauseTriggerInternal(TriggerKey triggerKey)
-    {
-        var trigger = await _triggerRepository.GetTrigger(triggerKey).ConfigureAwait(false);
-
-        var state = trigger?.State ?? Models.TriggerState.Deleted;
-        switch (state)
-        {
-            case Models.TriggerState.Waiting:
-            case Models.TriggerState.Acquired:
-            {
-                await _triggerRepository.UpdateTriggerState(triggerKey, Models.TriggerState.Paused)
-                    .ConfigureAwait(false);
-                break;
-            }
-            case Models.TriggerState.Blocked:
-            {
-                await _triggerRepository.UpdateTriggerState(triggerKey, Models.TriggerState.PausedBlocked)
-                    .ConfigureAwait(false);
-                break;
-            }
-        }
-    }
-
 
     private async Task<IReadOnlyCollection<string>> PauseTriggerGroupInternal(GroupMatcher<TriggerKey> matcher)
     {
@@ -1239,7 +716,6 @@ public class MongoDbJobStore : IJobStore
         return removedTrigger > 0;
     }
 
-
     private async Task<bool> RemoveJobInternal(JobKey jobKey)
     {
         await _triggerRepository.DeleteTriggers(jobKey).ConfigureAwait(false);
@@ -1247,91 +723,6 @@ public class MongoDbJobStore : IJobStore
         var result = await _jobDetailRepository.DeleteJob(jobKey).ConfigureAwait(false);
         return result > 0;
     }
-
-    private async Task<bool> RemoveTriggerInternal(TriggerKey key, IJobDetail? job = null)
-    {
-        var trigger = await _triggerRepository.GetTrigger(key).ConfigureAwait(false);
-        if (trigger == null)
-        {
-            return false;
-        }
-
-        if (job == null)
-        {
-            var result = await _jobDetailRepository.GetJob(trigger.JobKey).ConfigureAwait(false);
-            job = result?.GetJobDetail();
-        }
-
-        var removedTrigger = await _triggerRepository.DeleteTrigger(key).ConfigureAwait(false) > 0;
-
-        if (job != null && !job.Durable)
-        {
-            if (await _triggerRepository.GetCount(job.Key).ConfigureAwait(false) == 0)
-            {
-                if (await RemoveJobInternal(job.Key).ConfigureAwait(false))
-                {
-                    await _schedulerSignaler.NotifySchedulerListenersJobDeleted(job.Key).ConfigureAwait(false);
-                }
-            }
-        }
-
-        return removedTrigger;
-    }
-
-
-    private async Task<bool> RemoveCalendarInternal(string calendarName)
-    {
-        if (await _triggerRepository.CalendarIsReferenced(calendarName).ConfigureAwait(false))
-        {
-            throw new JobPersistenceException("Calender cannot be removed if it referenced by a trigger!");
-        }
-
-        return await _calendarRepository.DeleteCalendar(calendarName).ConfigureAwait(false) > 0;
-    }
-
-
-    private async Task ResumeTriggerInternal(TriggerKey triggerKey)
-    {
-        var trigger = await _triggerRepository.GetTrigger(triggerKey).ConfigureAwait(false);
-
-        if (trigger?.NextFireTime == null || trigger.NextFireTime == DateTime.MinValue)
-        {
-            return;
-        }
-
-        var blocked = trigger.State == Models.TriggerState.PausedBlocked;
-        var newState = await CheckBlockedState(trigger.JobKey, Models.TriggerState.Waiting).ConfigureAwait(false);
-        var misfired = false;
-
-        if (_schedulerRunning && trigger.NextFireTime < SystemTime.UtcNow())
-        {
-            misfired = await UpdateMisfiredTrigger(triggerKey, newState, true).ConfigureAwait(false);
-        }
-
-        if (!misfired)
-        {
-            var oldState = blocked ? Models.TriggerState.PausedBlocked : Models.TriggerState.Paused;
-
-            await _triggerRepository.UpdateTriggerState(triggerKey, newState, oldState).ConfigureAwait(false);
-        }
-    }
-
-
-    private async Task<IReadOnlyCollection<string>> ResumeTriggersInternal(GroupMatcher<TriggerKey> matcher)
-    {
-        await _pausedTriggerGroupRepository.DeletePausedTriggerGroup(matcher).ConfigureAwait(false);
-        var groups = new HashSet<string>();
-
-        var keys = await _triggerRepository.GetTriggerKeys(matcher).ConfigureAwait(false);
-        foreach (var triggerKey in keys)
-        {
-            await ResumeTriggerInternal(triggerKey).ConfigureAwait(false);
-            groups.Add(triggerKey.Group);
-        }
-
-        return groups.ToList();
-    }
-
 
     private async Task ResumeAllInternal()
     {
@@ -1344,52 +735,6 @@ public class MongoDbJobStore : IJobStore
 
         await _pausedTriggerGroupRepository.DeletePausedTriggerGroup(AllGroupsPaused).ConfigureAwait(false);
     }
-
-
-    private async Task StoreCalendarInternal(
-        string calName,
-        ICalendar calendar,
-        bool replaceExisting,
-        bool updateTriggers,
-        CancellationToken token = default
-    )
-    {
-        var existingCal = await CalendarExists(calName, token).ConfigureAwait(false);
-        if (existingCal && !replaceExisting)
-        {
-            throw new ObjectAlreadyExistsException($"Calendar with name '{calName}' already exists.");
-        }
-
-        var persistentCalendar = new Calendar(calName, calendar, InstanceName);
-
-        if (existingCal)
-        {
-            if (await _calendarRepository.UpdateCalendar(persistentCalendar).ConfigureAwait(false) < 1)
-            {
-                throw new JobPersistenceException("Couldn't store calendar.  Update failed.");
-            }
-
-            if (updateTriggers)
-            {
-                var triggers = await _triggerRepository.SelectTriggersForCalendar(calName).ConfigureAwait(false);
-
-                foreach (var trigger in triggers)
-                {
-                    var quartzTrigger = trigger.GetTrigger();
-
-                    quartzTrigger.UpdateWithNewCalendar(calendar, MisfireThreshold);
-
-                    await StoreTriggerInternal(quartzTrigger, null, true, Models.TriggerState.Waiting, false, false)
-                        .ConfigureAwait(false);
-                }
-            }
-        }
-        else
-        {
-            await _calendarRepository.AddCalendar(persistentCalendar).ConfigureAwait(false);
-        }
-    }
-
 
     private async Task StoreJobInternal(IJobDetail newJob, bool replaceExisting)
     {
@@ -1411,425 +756,6 @@ public class MongoDbJobStore : IJobStore
             await _jobDetailRepository.AddJob(jobDetail).ConfigureAwait(false);
         }
     }
-
-
-    private async Task StoreTriggerInternal(
-        IOperableTrigger newTrigger,
-        IJobDetail? job,
-        bool replaceExisting,
-        Models.TriggerState state,
-        bool forceState,
-        bool recovering
-    )
-    {
-        var existingTrigger = await _triggerRepository.TriggerExists(newTrigger.Key).ConfigureAwait(false);
-
-        if (existingTrigger && !replaceExisting)
-        {
-            throw new ObjectAlreadyExistsException(newTrigger);
-        }
-
-        if (!forceState)
-        {
-            var shouldBePaused = await _pausedTriggerGroupRepository.IsTriggerGroupPaused(newTrigger.Key.Group)
-                .ConfigureAwait(false);
-
-            if (!shouldBePaused)
-            {
-                shouldBePaused = await _pausedTriggerGroupRepository.IsTriggerGroupPaused(AllGroupsPaused)
-                    .ConfigureAwait(false);
-                if (shouldBePaused)
-                {
-                    await _pausedTriggerGroupRepository.AddPausedTriggerGroup(newTrigger.Key.Group)
-                        .ConfigureAwait(false);
-                }
-            }
-
-            if (shouldBePaused && (state == Models.TriggerState.Waiting || state == Models.TriggerState.Acquired))
-            {
-                state = Models.TriggerState.Paused;
-            }
-        }
-
-        if (job == null)
-        {
-            var jobDetail = await _jobDetailRepository.GetJob(newTrigger.JobKey).ConfigureAwait(false);
-            job = jobDetail?.GetJobDetail();
-        }
-
-        if (job == null)
-        {
-            throw new JobPersistenceException(
-                $"The job ({newTrigger.JobKey}) referenced by the trigger does not exist."
-            );
-        }
-
-        if (job.ConcurrentExecutionDisallowed && !recovering)
-        {
-            state = await CheckBlockedState(job.Key, state).ConfigureAwait(false);
-        }
-
-
-        var trigger = TriggerFactory.CreateTrigger(newTrigger, state, InstanceName);
-
-        if (existingTrigger)
-        {
-            await _triggerRepository.UpdateTrigger(trigger).ConfigureAwait(false);
-        }
-        else
-        {
-            await _triggerRepository.AddTrigger(trigger).ConfigureAwait(false);
-        }
-    }
-
-
-    private async Task<Models.TriggerState> CheckBlockedState(JobKey jobKey, Models.TriggerState currentState)
-    {
-        // State can only transition to BLOCKED from PAUSED or WAITING.
-        if (currentState != Models.TriggerState.Waiting && currentState != Models.TriggerState.Paused)
-        {
-            return currentState;
-        }
-
-        var firedTriggers = await _firedTriggerRepository.GetFiredTriggers(jobKey).ConfigureAwait(false);
-
-        var firedTrigger = firedTriggers.FirstOrDefault();
-        if (firedTrigger != null)
-        {
-            if (firedTrigger.ConcurrentExecutionDisallowed) // TODO: worry about failed/recovering/volatile job  states?
-            {
-                return currentState == Models.TriggerState.Paused ? Models.TriggerState.PausedBlocked
-                    : Models.TriggerState.Blocked;
-            }
-        }
-
-        return currentState;
-    }
-
-
-    private async Task<TriggerFiredBundle?> TriggerFiredInternal(IOperableTrigger trigger)
-    {
-        // Make sure trigger wasn't deleted, paused, or completed...
-        var state = await _triggerRepository.GetTriggerState(trigger.Key).ConfigureAwait(false);
-        if (state != Models.TriggerState.Acquired)
-        {
-            return null;
-        }
-
-        JobDetail? job;
-        try
-        {
-            job = await _jobDetailRepository.GetJob(trigger.JobKey).ConfigureAwait(false);
-            if (job == null)
-            {
-                return null;
-            }
-        }
-        catch (JobPersistenceException ex)
-        {
-            _logger.LogError(ex, "Error retrieving job, setting trigger state to ERROR.");
-
-            await _triggerRepository.UpdateTriggerState(trigger.Key, Models.TriggerState.Error).ConfigureAwait(false);
-            throw;
-        }
-
-        ICalendar? calendar = null;
-        if (trigger.CalendarName != null)
-        {
-            calendar = await _calendarRepository.GetCalendar(trigger.CalendarName).ConfigureAwait(false);
-            if (calendar == null)
-            {
-                return null;
-            }
-        }
-
-        await _firedTriggerRepository.UpdateFiredTrigger(
-                new FiredTrigger(
-                    trigger.FireInstanceId,
-                    TriggerFactory.CreateTrigger(trigger, Models.TriggerState.Executing, InstanceName),
-                    job
-                )
-                {
-                    InstanceId = InstanceId,
-                    State = Models.TriggerState.Executing,
-                }
-            )
-            .ConfigureAwait(false);
-
-        var prevFireTime = trigger.GetPreviousFireTimeUtc();
-
-        // call triggered - to update the trigger's next-fire-time state...
-        trigger.Triggered(calendar);
-
-        state = Models.TriggerState.Waiting;
-        var force = true;
-
-        if (job.ConcurrentExecutionDisallowed)
-        {
-            state = Models.TriggerState.Blocked;
-            force = false;
-            await _triggerRepository.UpdateTriggersStates(
-                    trigger.JobKey,
-                    Models.TriggerState.Blocked,
-                    Models.TriggerState.Waiting
-                )
-                .ConfigureAwait(false);
-            await _triggerRepository.UpdateTriggersStates(
-                    trigger.JobKey,
-                    Models.TriggerState.Blocked,
-                    Models.TriggerState.Acquired
-                )
-                .ConfigureAwait(false);
-            await _triggerRepository.UpdateTriggersStates(
-                    trigger.JobKey,
-                    Models.TriggerState.PausedBlocked,
-                    Models.TriggerState.Paused
-                )
-                .ConfigureAwait(false);
-        }
-
-        if (!trigger.GetNextFireTimeUtc().HasValue)
-        {
-            state = Models.TriggerState.Complete;
-            force = true;
-        }
-
-        var jobDetail = job.GetJobDetail();
-        await StoreTriggerInternal(trigger, jobDetail, true, state, force, force).ConfigureAwait(false);
-
-        jobDetail.JobDataMap.ClearDirtyFlag();
-
-        return new TriggerFiredBundle(
-            jobDetail,
-            trigger,
-            calendar,
-            trigger.Key.Group.Equals(SchedulerConstants.DefaultRecoveryGroup),
-            SystemTime.UtcNow(),
-            trigger.GetPreviousFireTimeUtc(),
-            prevFireTime,
-            trigger.GetNextFireTimeUtc()
-        );
-    }
-
-
-    private async Task<bool> UpdateMisfiredTrigger(
-        TriggerKey triggerKey,
-        Models.TriggerState newStateIfNotComplete,
-        bool forceState
-    )
-    {
-        var trigger = await _triggerRepository.GetTrigger(triggerKey).ConfigureAwait(false);
-        if (trigger == null)
-        {
-            throw new JobPersistenceException($"No trigger found with id {triggerKey}");
-        }
-
-        var misfireTime = SystemTime.UtcNow();
-        if (MisfireThreshold > TimeSpan.Zero)
-        {
-            misfireTime = misfireTime.AddMilliseconds(-1 * MisfireThreshold.TotalMilliseconds);
-        }
-
-        if (trigger.NextFireTime.GetValueOrDefault() > misfireTime)
-        {
-            return false;
-        }
-
-        await DoUpdateOfMisfiredTrigger(trigger, forceState, newStateIfNotComplete, false).ConfigureAwait(false);
-
-        return true;
-    }
-
-
-    private async Task DoUpdateOfMisfiredTrigger(
-        Trigger trigger,
-        bool forceState,
-        Models.TriggerState newStateIfNotComplete,
-        bool recovering
-    )
-    {
-        var operableTrigger = trigger.GetTrigger();
-
-        ICalendar? cal = null;
-        if (trigger.CalendarName != null)
-        {
-            cal = await _calendarRepository.GetCalendar(trigger.CalendarName).ConfigureAwait(false);
-        }
-
-        await _schedulerSignaler.NotifyTriggerListenersMisfired(operableTrigger).ConfigureAwait(false);
-        operableTrigger.UpdateAfterMisfire(cal);
-
-        if (!operableTrigger.GetNextFireTimeUtc().HasValue)
-        {
-            await StoreTriggerInternal(
-                    operableTrigger,
-                    null,
-                    true,
-                    Models.TriggerState.Complete,
-                    forceState,
-                    recovering
-                )
-                .ConfigureAwait(false);
-            await _schedulerSignaler.NotifySchedulerListenersFinalized(operableTrigger).ConfigureAwait(false);
-        }
-        else
-        {
-            await StoreTriggerInternal(operableTrigger, null, true, newStateIfNotComplete, forceState, recovering)
-                .ConfigureAwait(false);
-        }
-    }
-
-
-    private async Task<IReadOnlyCollection<IOperableTrigger>> AcquireNextTriggersInternal(
-        DateTimeOffset noLaterThan,
-        int maxCount,
-        TimeSpan timeWindow
-    )
-    {
-        ArgumentOutOfRangeException.ThrowIfLessThan(timeWindow, TimeSpan.Zero);
-
-        var acquiredTriggers = new List<IOperableTrigger>();
-        var acquiredJobKeysForNoConcurrentExec = new HashSet<JobKey>();
-
-        const int maxDoLoopRetry = 3;
-        var currentLoopCount = 0;
-
-        do
-        {
-            currentLoopCount++;
-
-            var triggerKeys = await _triggerRepository
-                .GetTriggersToAcquire(noLaterThan + timeWindow, MisfireTime, maxCount)
-                .ConfigureAwait(false);
-
-            if (triggerKeys.Count == 0)
-            {
-                return acquiredTriggers;
-            }
-
-
-            var batchEnd = noLaterThan;
-
-            foreach (var triggerKey in triggerKeys)
-            {
-                // If our trigger is no longer available, try a new one.
-                var nextTrigger = await _triggerRepository.GetTrigger(triggerKey).ConfigureAwait(false);
-                if (nextTrigger == null)
-                {
-                    continue; // next trigger
-                }
-
-                // In the Ado version we'd join the TRIGGERS with JOB_DETAILS. 
-                var jobKey = nextTrigger.JobKey;
-
-                JobDetail? jobDetail;
-                try
-                {
-                    jobDetail = await _jobDetailRepository.GetJob(jobKey).ConfigureAwait(false);
-                    if (jobDetail == null)
-                    {
-                        throw new JobPersistenceException($"No job details found for trigger {triggerKey}");
-                    }
-                }
-                catch (Exception)
-                {
-                    await _triggerRepository.UpdateTriggerState(triggerKey, Models.TriggerState.Error)
-                        .ConfigureAwait(false);
-                    continue;
-                }
-
-                // If trigger's job is set as @DisallowConcurrentExecution, and it has already been added to result, then
-                // put it back into the timeTriggers set and continue to search for next trigger.
-
-                if (jobDetail.ConcurrentExecutionDisallowed)
-                {
-                    if (!acquiredJobKeysForNoConcurrentExec.Add(jobKey))
-                    {
-                        continue; // next trigger
-                    }
-                }
-
-
-                var nextFireTimeUtc = nextTrigger.NextFireTime;
-
-                // A trigger should not return NULL on nextFireTime when fetched from DB.
-                // But for whatever reason if we do have this (BAD trigger implementation or
-                // data?), we then should log a warning and continue to next trigger.
-                // User would need to manually fix these triggers from DB as they will not
-                // able to be clean up by Quartz since we are not returning it to be processed.
-                if (nextFireTimeUtc == null)
-                {
-                    _logger.LogWarning(
-                        "Trigger {Key} returned null on nextFireTime and yet still exists in DB!",
-                        nextTrigger.GetTriggerKey()
-                    );
-                    continue;
-                }
-
-                if (nextFireTimeUtc > batchEnd)
-                {
-                    break;
-                }
-
-                // We now have a acquired trigger, let's add to return list.
-                // If our trigger was no longer in the expected state, try a new one.
-                var result = await _triggerRepository.UpdateTriggerState(
-                        triggerKey,
-                        Models.TriggerState.Acquired,
-                        Models.TriggerState.Waiting
-                    )
-                    .ConfigureAwait(false);
-
-                if (result <= 0)
-                {
-                    continue;
-                }
-
-                var operableTrigger = nextTrigger.GetTrigger();
-                operableTrigger.FireInstanceId = GetFiredTriggerRecordId();
-
-                await _firedTriggerRepository.AddFiredTrigger(
-                        new FiredTrigger(operableTrigger.FireInstanceId, nextTrigger, null)
-                        {
-                            State = Models.TriggerState.Acquired,
-                            InstanceId = InstanceId,
-                        }
-                    )
-                    .ConfigureAwait(false);
-
-                if (acquiredTriggers.Count == 0)
-                {
-                    var now = SystemTime.UtcNow();
-                    var nextFireTime = nextFireTimeUtc.Value;
-                    var max = now > nextFireTime ? now : nextFireTime;
-
-                    batchEnd = max + timeWindow;
-                }
-
-                acquiredTriggers.Add(operableTrigger);
-            }
-
-            // if we didn't end up with any trigger to fire from that first
-            // batch, try again for another batch. We allow with a max retry count.
-
-            if (acquiredTriggers.Count == 0 && currentLoopCount < maxDoLoopRetry)
-            {
-                continue;
-            }
-
-            // We are done with the while loop.
-            break;
-        } while (true);
-
-        return acquiredTriggers;
-    }
-
-    private string GetFiredTriggerRecordId()
-    {
-        Interlocked.Increment(ref _fireTriggerRecordCounter);
-        return InstanceId + _fireTriggerRecordCounter;
-    }
-
 
     private async Task TriggeredJobCompleteInternal(
         IOperableTrigger trigger,
@@ -1958,534 +884,6 @@ public class MongoDbJobStore : IJobStore
     {
         _schedulerSignaler.SignalSchedulingChange(candidateNewNextFireTime);
     }
-
-
-    private async Task RecoverJobsInternal()
-    {
-        // update inconsistent job states
-        var result = await _triggerRepository.UpdateTriggersStates(
-                Models.TriggerState.Waiting,
-                Models.TriggerState.Acquired,
-                Models.TriggerState.Blocked
-            )
-            .ConfigureAwait(false);
-
-        result += await _triggerRepository
-            .UpdateTriggersStates(Models.TriggerState.Paused, Models.TriggerState.PausedBlocked)
-            .ConfigureAwait(false);
-
-        _logger.LogInformation("Freed {Count} triggers from 'acquired' / 'blocked' state.", result);
-
-        // clean up misfired jobs
-        await RecoverMisfiredJobsInternal(true).ConfigureAwait(false);
-
-        var results =
-            (await _firedTriggerRepository.GetRecoverableFiredTriggers(InstanceId).ConfigureAwait(false)).Select(async
-                trigger => trigger.GetRecoveryTrigger(
-                    await _triggerRepository.GetTriggerJobDataMap(trigger.TriggerKey).ConfigureAwait(false)
-                )
-            );
-
-        // recover jobs marked for recovery that were not fully executed
-        var recoveringJobTriggers = (await Task.WhenAll(results).ConfigureAwait(false)).ToList();
-
-        _logger.LogInformation(
-            "Recovering {Count} jobs that were in-progress at the time of the last shut-down.",
-            recoveringJobTriggers.Count
-        );
-
-        foreach (var recoveringJobTrigger in recoveringJobTriggers)
-        {
-            if (await _jobDetailRepository.JobExists(recoveringJobTrigger.JobKey).ConfigureAwait(false))
-            {
-                recoveringJobTrigger.ComputeFirstFireTimeUtc(null);
-
-                await StoreTriggerInternal(recoveringJobTrigger, null, false, Models.TriggerState.Waiting, false, true)
-                    .ConfigureAwait(false);
-            }
-        }
-
-        _logger.LogInformation("Recovery complete");
-
-        var completedTriggers =
-            await _triggerRepository.GetTriggerKeys(Models.TriggerState.Complete).ConfigureAwait(false);
-
-        foreach (var completedTrigger in completedTriggers)
-        {
-            await RemoveTriggerInternal(completedTrigger).ConfigureAwait(false);
-        }
-
-        _logger.LogInformation("Removed {Count} 'complete' triggers.", completedTriggers.Count);
-
-        // clean up any fired trigger entries
-        result = await _firedTriggerRepository.DeleteFiredTriggers().ConfigureAwait(false);
-        _logger.LogInformation("Removed {Count} stale fired job entries.", result);
-    }
-
-
-    private async Task<RecoverMisfiredJobsResult> RecoverMisfiredJobsInternal(bool recovering)
-    {
-        // If recovering, we want to handle all of the misfired
-        // triggers right away.
-        var maxMisfiresToHandleAtTime = recovering ? -1 : MaxMisfiresToHandleAtATime;
-
-        var (hasMoreMisfiredTriggers, misfiredTriggers) = await _triggerRepository.HasMisfiredTriggers(
-                MisfireTime,
-                maxMisfiresToHandleAtTime
-            )
-            .ConfigureAwait(false);
-
-        if (hasMoreMisfiredTriggers)
-        {
-            _logger.LogInformation(
-                "Handling the first {Count} triggers that missed their scheduled fire-time.  " +
-                "More misfired triggers remain to be processed.",
-                misfiredTriggers.Count
-            );
-        }
-        else if (misfiredTriggers.Count > 0)
-        {
-            _logger.LogInformation(
-                "Handling {Count} trigger(s) that missed their scheduled fire-time.",
-                misfiredTriggers.Count
-            );
-        }
-        else
-        {
-            _logger.LogInformation("Found 0 triggers that missed their scheduled fire-time.");
-            return RecoverMisfiredJobsResult.NoOp;
-        }
-
-        var earliestNewTime = DateTimeOffset.MaxValue;
-
-        foreach (var misfiredTrigger in misfiredTriggers)
-        {
-            var trigger = await _triggerRepository.GetTrigger(misfiredTrigger).ConfigureAwait(false);
-            if (trigger == null)
-            {
-                continue;
-            }
-
-            await DoUpdateOfMisfiredTrigger(trigger, false, Models.TriggerState.Waiting, recovering)
-                .ConfigureAwait(false);
-
-            var nextTime = trigger.NextFireTime;
-            if (nextTime.HasValue && nextTime.Value < earliestNewTime)
-            {
-                earliestNewTime = nextTime.Value;
-            }
-        }
-
-        return new RecoverMisfiredJobsResult(hasMoreMisfiredTriggers, misfiredTriggers.Count, earliestNewTime);
-    }
-
-
-    #region Cluster
-
-    internal async Task<bool> DoCheckIn(CancellationToken cancellationToken = default)
-    {
-        var recovered = false;
-
-
-        try
-        {
-            await _pendingLocksSemaphore.WaitAsync(cancellationToken).ConfigureAwait(false);
-
-            await using var context = await _lockingManager.CreateLockContext(cancellationToken).ConfigureAwait(false);
-
-            try
-            {
-                // Other than the first time, always checkin first to make sure there is
-                // work to be done before we acquire the lock (since that is expensive,
-                // and is almost never necessary).  This must be done in a separate
-                // transaction to prevent a deadlock under recovery conditions.
-                IReadOnlyList<Scheduler>? failedRecords = null;
-                if (!_firstCheckIn)
-                {
-                    failedRecords = await ClusterCheckIn().ConfigureAwait(false);
-                }
-
-                if (_firstCheckIn || failedRecords != null && failedRecords.Count > 0)
-                {
-                    await context.TryAcquireLock(InstanceName, QuartzLockType.StateAccess, cancellationToken)
-                        .ConfigureAwait(false);
-
-
-                    // Now that we own the lock, make sure we still have work to do.
-                    // The first time through, we also need to make sure we update/create our state record
-                    if (_firstCheckIn)
-                    {
-                        failedRecords = await ClusterCheckIn().ConfigureAwait(false);
-                    }
-                    else
-                    {
-                        failedRecords = await FindFailedInstances().ConfigureAwait(false);
-                    }
-
-                    if (failedRecords.Count > 0)
-                    {
-                        await context.TryAcquireLock(InstanceName, QuartzLockType.TriggerAccess, cancellationToken)
-                            .ConfigureAwait(false);
-
-                        await ClusterRecover(failedRecords).ConfigureAwait(false);
-                        recovered = true;
-                    }
-                }
-
-                await context.CommitTransaction(cancellationToken).ConfigureAwait(false);
-            }
-            catch (Exception ex)
-            {
-                await context.RollbackTransaction(cancellationToken).ConfigureAwait(false);
-
-                _logger.LogWarning(ex, "Failed to do cluster check-in {Message}", ex.Message);
-            }
-        }
-        finally
-        {
-            _pendingLocksSemaphore.Release();
-        }
-
-        _firstCheckIn = false;
-
-        return recovered;
-    }
-
-    /// <summary>
-    /// Get a list of all scheduler instances in the cluster that may have failed.
-    /// This includes this scheduler if it is checking in for the first time.
-    /// </summary>
-    private async Task<List<Scheduler>> FindFailedInstances()
-    {
-        var failedInstances = new List<Scheduler>();
-        var foundThisScheduler = false;
-
-        var schedulers = await _schedulerRepository.SelectSchedulerStateRecords(null).ConfigureAwait(false);
-
-        foreach (var scheduler in schedulers)
-        {
-            // find own record...
-            if (scheduler.InstanceId == InstanceId)
-            {
-                foundThisScheduler = true;
-                if (_firstCheckIn)
-                {
-                    failedInstances.Add(scheduler);
-                }
-            }
-            else
-            {
-                // find failed instances...
-                if (CalcFailedIfAfter(scheduler) < DateTimeOffset.UtcNow)
-                {
-                    failedInstances.Add(scheduler);
-                }
-            }
-        }
-
-        // The first time through, also check for orphaned fired triggers.
-        if (_firstCheckIn)
-        {
-            var orphanedInstances = await FindOrphanedFailedInstances(schedulers).ConfigureAwait(false);
-            failedInstances.AddRange(orphanedInstances);
-        }
-
-        // If not the first time but we didn't find our own instance, then
-        // Someone must have done recovery for us.
-        if (!foundThisScheduler && !_firstCheckIn)
-        {
-            // TODO: revisit when handle self-failed-out impl'ed (see TODO in clusterCheckIn() below)
-            _logger.LogWarning(
-                "This scheduler instance ({InstanceId}) is still active but was recovered by another instance in the cluster. This may cause inconsistent behavior.",
-                InstanceId
-            );
-        }
-
-        return failedInstances;
-    }
-
-    /// <summary>
-    /// Create dummy <see cref="SchedulerStateRecord" /> objects for fired triggers
-    /// that have no scheduler state record. Checkin timestamp and interval are
-    /// left as zero on these dummy <see cref="SchedulerStateRecord" /> objects.
-    /// </summary>
-    /// <param name="schedulers">List of all current <see cref="SchedulerStateRecord" />s</param>
-    private async Task<IReadOnlyList<Scheduler>> FindOrphanedFailedInstances(IReadOnlyCollection<Scheduler> schedulers)
-    {
-        var orphanedInstances = new List<Scheduler>();
-
-        var ids = await _firedTriggerRepository.SelectFiredTriggerInstanceIds().ConfigureAwait(false);
-
-        var allFiredTriggerInstanceIds = new HashSet<string>(ids);
-        if (allFiredTriggerInstanceIds.Count > 0)
-        {
-            foreach (var scheduler in schedulers)
-            {
-                allFiredTriggerInstanceIds.Remove(scheduler.InstanceId);
-            }
-
-            foreach (var instanceId in allFiredTriggerInstanceIds)
-            {
-                var orphanedInstance = new Scheduler(InstanceName, instanceId);
-                orphanedInstances.Add(orphanedInstance);
-
-                _logger.LogWarning(
-                    "Found orphaned fired triggers for instance: {SchedulerName}",
-                    orphanedInstance.SchedulerName
-                );
-            }
-        }
-
-        return orphanedInstances;
-    }
-
-
-    private DateTimeOffset CalcFailedIfAfter(Scheduler scheduler)
-    {
-        var passed = DateTimeOffset.UtcNow - LastCheckin;
-        var ts = scheduler.CheckInInterval > passed ? scheduler.CheckInInterval : passed; // Max
-
-        return scheduler.LastCheckIn
-            //
-            .Add(ts)
-            .Add(ClusterCheckinMisfireThreshold);
-    }
-
-    private async Task<IReadOnlyList<Scheduler>> ClusterCheckIn()
-    {
-        var failedInstances = await FindFailedInstances().ConfigureAwait(false);
-
-        try
-        {
-            // TODO: handle self-failed-out
-
-            // check in...
-            LastCheckin = SystemTime.UtcNow();
-
-            if (await _schedulerRepository.UpdateState(InstanceId, LastCheckin).ConfigureAwait(false) == 0)
-            {
-                await _schedulerRepository.AddScheduler(InstanceId, LastCheckin, ClusterCheckinInterval)
-                    .ConfigureAwait(false);
-            }
-        }
-        catch (Exception e)
-        {
-            throw new JobPersistenceException("Failure updating scheduler state when checking-in: " + e.Message, e);
-        }
-
-        return failedInstances;
-    }
-
-    private async Task ClusterRecover(IReadOnlyList<Scheduler> failedInstances)
-    {
-        if (failedInstances.Count <= 0)
-        {
-            return;
-        }
-
-        var recoverIds = SystemTime.UtcNow().Ticks;
-
-        LogWarnIfNonZero(
-            failedInstances.Count,
-            "ClusterManager: detected {Count} failed or restarted instances.",
-            failedInstances.Count
-        );
-
-        try
-        {
-            foreach (var rec in failedInstances)
-            {
-                _logger.LogInformation(
-                    "ClusterManager: Scanning for instance \"{InstanceId}\"'s failed in-progress jobs.",
-                    rec.InstanceId
-                );
-
-
-                var firedTriggerRecs = await _firedTriggerRepository.SelectInstancesFiredTriggerRecords(rec.InstanceId)
-                    .ConfigureAwait(false);
-
-                var acquiredCount = 0;
-                var recoveredCount = 0;
-                var otherCount = 0;
-
-                var triggerKeys = new HashSet<TriggerKey>();
-
-                foreach (var ftRec in firedTriggerRecs)
-                {
-                    var tKey = ftRec.TriggerKey;
-                    var jKey = ftRec.JobKey;
-
-                    triggerKeys.Add(tKey);
-
-                    switch (ftRec.State)
-                    {
-                        // release blocked triggers..
-                        case Models.TriggerState.Blocked:
-                        {
-                            await _triggerRepository.UpdateTriggersStates(
-                                    jKey,
-                                    Models.TriggerState.Waiting,
-                                    Models.TriggerState.Blocked
-                                )
-                                .ConfigureAwait(false);
-                            break;
-                        }
-                        case Models.TriggerState.PausedBlocked:
-                        {
-                            await _triggerRepository.UpdateTriggersStates(
-                                    jKey,
-                                    Models.TriggerState.Paused,
-                                    Models.TriggerState.PausedBlocked
-                                )
-                                .ConfigureAwait(false);
-                            break;
-                        }
-                    }
-
-                    // release acquired triggers..
-                    if (ftRec.State == Models.TriggerState.Acquired)
-                    {
-                        await _triggerRepository.UpdateTriggerState(
-                                tKey,
-                                Models.TriggerState.Waiting,
-                                Models.TriggerState.Acquired
-                            )
-                            .ConfigureAwait(false);
-                        acquiredCount++;
-                    }
-                    else if (ftRec.RequestsRecovery)
-                    {
-                        // handle jobs marked for recovery that were not fully
-                        // executed...
-                        if (await _jobDetailRepository.JobExists(jKey).ConfigureAwait(false))
-                        {
-                            var recoveryTrig = new SimpleTriggerImpl(
-                                $"recover_{rec.InstanceId}_{Convert.ToString(recoverIds++, CultureInfo.InvariantCulture)}",
-                                SchedulerConstants.DefaultRecoveryGroup,
-                                ftRec.Fired
-                            )
-                            {
-                                JobName = jKey.Name,
-                                JobGroup = jKey.Group,
-                                MisfireInstruction = MisfireInstruction.SimpleTrigger.FireNow,
-                                Priority = ftRec.Priority,
-                            };
-
-                            var jd = await _triggerRepository.GetTriggerJobDataMap(tKey).ConfigureAwait(false);
-
-                            jd.Put(SchedulerConstants.FailedJobOriginalTriggerName, tKey.Name);
-                            jd.Put(SchedulerConstants.FailedJobOriginalTriggerGroup, tKey.Group);
-                            jd.Put(
-                                SchedulerConstants.FailedJobOriginalTriggerFiretime,
-                                Convert.ToString(ftRec.Fired, CultureInfo.InvariantCulture)
-                            );
-                            recoveryTrig.JobDataMap = jd;
-
-                            recoveryTrig.ComputeFirstFireTimeUtc(null);
-                            await StoreTriggerInternal(
-                                    recoveryTrig,
-                                    null,
-                                    false,
-                                    Models.TriggerState.Waiting,
-                                    false,
-                                    true
-                                )
-                                .ConfigureAwait(false);
-                            recoveredCount++;
-                        }
-                        else
-                        {
-                            _logger.LogWarning(
-                                "ClusterManager: failed job '{Key}' no longer exists, cannot schedule recovery.",
-                                jKey
-                            );
-                            otherCount++;
-                        }
-                    }
-                    else
-                    {
-                        otherCount++;
-                    }
-
-                    // free up stateful job's triggers
-                    if (ftRec.ConcurrentExecutionDisallowed)
-                    {
-                        await _triggerRepository.UpdateTriggersStates(
-                                jKey,
-                                Models.TriggerState.Waiting,
-                                Models.TriggerState.Blocked
-                            )
-                            .ConfigureAwait(false);
-                        await _triggerRepository.UpdateTriggersStates(
-                                jKey,
-                                Models.TriggerState.Paused,
-                                Models.TriggerState.PausedBlocked
-                            )
-                            .ConfigureAwait(false);
-                    }
-                }
-
-                await _firedTriggerRepository.DeleteFiredTriggersByInstanceId(rec.InstanceId).ConfigureAwait(false);
-
-                // Check if any of the fired triggers we just deleted were the last fired trigger
-                // records of a COMPLETE trigger.
-                var completeCount = 0;
-                foreach (var triggerKey in triggerKeys)
-                {
-                    var triggerState = await _triggerRepository.GetTriggerState(triggerKey).ConfigureAwait(false);
-                    if (triggerState == Models.TriggerState.Complete)
-                    {
-                        var firedTriggers = await _firedTriggerRepository.SelectFiredTriggerRecords(
-                                triggerKey.Name,
-                                triggerKey.Group
-                            )
-                            .ConfigureAwait(false);
-
-                        if (firedTriggers.Count == 0)
-                        {
-                            if (await RemoveTriggerInternal(triggerKey).ConfigureAwait(false))
-                            {
-                                completeCount++;
-                            }
-                        }
-                    }
-                }
-
-                LogWarnIfNonZero(
-                    acquiredCount,
-                    "ClusterManager: ......Freed {Acquired} acquired trigger(s).",
-                    acquiredCount
-                );
-                LogWarnIfNonZero(
-                    completeCount,
-                    "ClusterManager: ......Deleted {Completed} complete triggers(s).",
-                    completeCount
-                );
-                LogWarnIfNonZero(
-                    recoveredCount,
-                    "ClusterManager: ......Scheduled {Recovered} recoverable job(s) for recovery.",
-                    recoveredCount
-                );
-                LogWarnIfNonZero(
-                    otherCount,
-                    "ClusterManager: ......Cleaned-up {Other} other failed job(s).",
-                    otherCount
-                );
-
-                if (rec.InstanceId.Equals(InstanceId) == false)
-                {
-                    await _schedulerRepository.DeleteScheduler(rec.InstanceId).ConfigureAwait(false);
-                }
-            }
-        }
-        catch (Exception e)
-        {
-            throw new JobPersistenceException("Failure recovering jobs: " + e.Message, e);
-        }
-    }
-
-    #endregion
-
 
     private void LogWarnIfNonZero(int val, [StructuredMessageTemplate] string? message, params object?[] args)
     {
