+++ src/Reddoxx.Quartz.MongoDbJobStore/Repositories/PausedTriggerGroupRepository.cs
@@ -44,11 +44,10 @@ internal class PausedTriggerGroupRepository : BaseRepository<PausedTriggerGroup>
         var filter = FilterBuilder.Eq(x => x.InstanceName, InstanceName);
 
         return await Collection
-            //
-            .Find(filter)
-            .Project(group => group.Group)
-            .ToListAsync()
-            .ConfigureAwait(false);
+                     //
+                     .Find(filter)
+                     .Project(group => group.Group)
+                     .ToListAsync();
     }
 
     public async Task<bool> IsTriggerGroupPaused(string group)
@@ -57,24 +56,21 @@ internal class PausedTriggerGroupRepository : BaseRepository<PausedTriggerGroup>
         var filter = FilterBuilder.Eq(x => x.InstanceName, InstanceName) & //
                      FilterBuilder.Eq(x => x.Group, group);
 
-        return await Collection
-            //
-            .Find(filter)
-            .AnyAsync()
-            .ConfigureAwait(false);
+        return await Collection.Find(filter)
+                               .AnyAsync();
     }
 
     public async Task AddPausedTriggerGroup(string group)
     {
         // INSERT INTO PAUSED_TRIGGER_GRPS (SCHED_NAME, TRIGGER_GROUP) VALUES (@schedulerName, @triggerGroup)
-        await Collection.InsertOneAsync(
-                new PausedTriggerGroup
-                {
-                    InstanceName = InstanceName,
-                    Group = group,
-                }
-            )
-            .ConfigureAwait(false);
+
+        var triggerGroup = new PausedTriggerGroup
+        {
+            InstanceName = InstanceName,
+            Group = group,
+        };
+
+        await Collection.InsertOneAsync(triggerGroup);
     }
 
     public async Task DeletePausedTriggerGroup(GroupMatcher<TriggerKey> matcher)
@@ -83,7 +79,7 @@ internal class PausedTriggerGroupRepository : BaseRepository<PausedTriggerGroup>
         var filter = FilterBuilder.Eq(x => x.InstanceName, InstanceName) &
                      FilterBuilder.Regex(x => x.Group, matcher.ToBsonRegularExpression());
 
-        await Collection.DeleteManyAsync(filter).ConfigureAwait(false);
+        await Collection.DeleteManyAsync(filter);
     }
 
     public async Task DeletePausedTriggerGroup(string groupName)
@@ -92,6 +88,6 @@ internal class PausedTriggerGroupRepository : BaseRepository<PausedTriggerGroup>
         var filter = FilterBuilder.Eq(x => x.InstanceName, InstanceName) & // 
                      FilterBuilder.Eq(x => x.Group, groupName);
 
-        await Collection.DeleteOneAsync(filter).ConfigureAwait(false);
+        await Collection.DeleteOneAsync(filter);
     }
 }
